+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/package-info.java
@@ -30,11 +30,11 @@
  * <li>{@link is.codion.swing.framework.ui.EntityApplicationPanel#SQL_TRACING}
  * <li>{@link is.codion.swing.framework.ui.EntityEditorPanel#VALID_INDICATOR}
  * <li>{@link is.codion.swing.framework.ui.EntityEditorPanel#MODIFIED_INDICATOR}
+ * <li>{@link is.codion.swing.framework.ui.EntityEditorPanel#MODIFIED_WARNING}
  * <li>{@link is.codion.swing.framework.ui.EntityEditorPanel#DEFAULT_TEXT_FIELD_COLUMNS}
  * <li>{@link is.codion.swing.framework.ui.EntityEditPanel.Config#USE_SAVE_CAPTION}
  * <li>{@link is.codion.swing.framework.ui.EntityEditPanel.Config#USE_FOCUS_ACTIVATION}
  * <li>{@link is.codion.swing.framework.ui.EntityEditPanel.Config#INCLUDE_ENTITY_VIEWER}
- * <li>{@link is.codion.swing.framework.ui.EntityEditPanel.Config#MODIFIED_WARNING}
  * <li>{@link is.codion.swing.framework.ui.EntityPanel.Config#KEYBOARD_NAVIGATION}
  * <li>{@link is.codion.swing.framework.ui.EntityPanel.Config#DISPOSE_EDIT_DIALOG_ON_ESCAPE}
  * <li>{@link is.codion.swing.framework.ui.EntityPanel.Config#INCLUDE_TOGGLE_EDIT_PANEL_CONTROL}